+++ index.html
@@ -57,6 +57,6 @@
 		<div class="wall-right"></div>
 	</div>
 
-	<script src="./game.js" type="module"></script>
+	<script src="./index.js" type="module"></script>
 </body>
 </html>
