+++ index.js
@@ -1,0 +1,5 @@
+import Game from './gameClass.js';
+
+const game = new Game();
+
+game.start();
