--- main.js
@@ -1,5 +1,0 @@
-// import Game from './game.js';
-
-// const game = new Game();
-
-// game.start();
